+++ engines/support/config/initializers/account_lifecycle.rb
@@ -1,3 +1,5 @@
+require 'account_extension/tickets'
+
 ActiveSupport.on_load(:account) do
   include AccountExtension::Tickets
 end
