+++ lib/email.rb
